+++ CHANGES.md
@@ -1,6 +1,12 @@
 ﻿### 0.26.xx (202x-xx-xx xx:xx:00 UTC)
 
 * Add support for scene episode number to multiple episodes
+* Add TMDB TV info source
+* Change return unknown.png flag if a requested flag image file doesn't exist
+* Change use regular cast from season as show cast.
+* Change prefer recent cast in cast order.
+* Change filter non main appearance of main cast persons.
+* Add direct match to relative sort as first priority of search results.
 
 
 ### 0.25.25 (2022-01-18 14:55:00 UTC)